flatten packets
